+++ jgclark.Summaries/src/todayProgress.js
@@ -2,7 +2,7 @@
 //-----------------------------------------------------------------------------
 // Progress update for Today only
 // Jonathan Clark, @jgclark
-// Last updated 30.3.2024 for v0.20.1+, @jgclark
+// Last updated 2025-10-07 for v1.0.0, @jgclark
 //-----------------------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
@@ -13,21 +13,11 @@ import {
   type OccurrencesToLookFor,
   type SummariesConfig,
 } from './summaryHelpers'
-import {
-  // hyphenatedDate,
-  todaysDateISOString,
-  // YYYYMMDDDateStringFromDate,
-  // withinDateRange
-} from '@helpers/dateTime'
+import { todaysDateISOString } from '@helpers/dateTime'
 import { toNPLocaleDateString, } from '@helpers/NPdateTime'
+import { clo, logDebug, logError, logInfo, logWarn, timer } from '@helpers/dev'
 import {
-  clo, logDebug, logError, logInfo, logWarn, timer,
-  // overrideSettingsWithEncodedTypedArgs,
-} from '@helpers/dev'
-import {
-  // CaseInsensitiveMap,
   createPrettyRunPluginLink,
-  // createRunPluginCallbackUrl, displayTitle,
   formatWithFields,
   getTagParamsFromString,
 } from '@helpers/general'
@@ -44,7 +34,6 @@ export async function todayProgressFromTemplate(params: string = ''): Promise<st
   try {
     logDebug(pluginJson, `todayProgressFromTemplate() starting with params '${params}' (type: ${typeof params})`)
 
-    // let configFromParams = JSON.parse(params)
     const config = await getSummariesSettings()
     const heading = await getTagParamsFromString(params, 'todayProgressHeading', config.todayProgressHeading)
     const itemsToShowStr = await getTagParamsFromString(params, 'todayProgressItems', config.todayProgressItems)
@@ -65,16 +54,16 @@ export async function todayProgressFromTemplate(params: string = ''): Promise<st
 
 /**
  * This is the entry point for /command or x-callback use of makeProgressUpdate
- * @param {?string} itemsToShowArg as comma-separated list of @mentions or #tags
+ * @param {?string} itemsToShowArg as comma-separated list of @mentions or #tags (optional). If not provided, then the default items in setting 'todayProgressItems' will be used.
  * @param {?string} headingArg
  */
 export async function todayProgress(itemsToShowArg?: string, headingArg?: string): Promise<void> {
   try {
-    logDebug(pluginJson, `makeTodayProgress() starting with itemsToShowArg '${itemsToShowArg ? itemsToShowArg : '-'}'`)
+    logDebug(pluginJson, `todayProgress() starting with itemsToShowArg '${itemsToShowArg ? itemsToShowArg : '-'}'`)
     const itemsToShowArr: Array<string> = itemsToShowArg
       ? itemsToShowArg.split(',')
       : []
-    logDebug('makeTodayProgress()', `itemsToShowArr '${String(itemsToShowArr)}'`)
+    logDebug('todayProgress()', `itemsToShowArr '${String(itemsToShowArr)}'`)
 
     if (typeof headingArg === 'string') {
       const _summaryStr: string = await makeTodayProgress(itemsToShowArr, 'command', headingArg)
@@ -84,7 +73,7 @@ export async function todayProgress(itemsToShowArg?: string, headingArg?: string
     // NB: don't need to do anything with output
   }
   catch (err) {
-    logError(pluginJson, 'makeTodayProgress() ' + err.message)
+    logError('todayProgress', `Error: ${err.message}`)
     return // for completeness
   }
 }
@@ -95,12 +84,16 @@ export async function todayProgress(itemsToShowArg?: string, headingArg?: string
  * If it's week to date, then use the user's first day of week from NP setting.
  * @author @jgclark
  *
- * @param {?string} itemsToShowArr e.g. ['@calories', '#test']
- * @param {?string} source of this call (command/xcb/template)
- * @param {?string} headingArg
+ * @param {?string} itemsToShowArr e.g. ['@calories', '#test']. If not provided, then the default items in setting 'todayProgressItems' will be used instead.
+ * @param {?string} source of this call (command/xcb/template). If not provided, then it assumes this is a 'command' call.
+ * @param {?string} headingArg (optional)
  * @returns {?string} - either return string to Template, or void to plugin
  */
-export async function makeTodayProgress(itemsToShowArr: Array<string> = [], source: string = 'command', headingArg?: string): Promise<string> {
+export async function makeTodayProgress(
+  itemsToShowArr: Array<string> = [],
+  source: string = 'command',
+  headingArg?: string
+): Promise<string> {
   try {
     // Get config setting
     const config: SummariesConfig = await getSummariesSettings()
@@ -124,14 +117,12 @@ export async function makeTodayProgress(itemsToShowArr: Array<string> = [], sour
 
     const settingsForGO: OccurrencesToLookFor = {
       GOYesNo: [],
-      GOHashtagsCount: [],
+      GOHashtagsCount: [], // covered in the total
       GOHashtagsTotal: (hashtagsToShow.length > 0) ? hashtagsToShow : [],
-      GOHashtagsAverage: [],
-      GOHashtagsExclude: [],
-      GOMentionsCount: [],
+      GOHashtagsAverage: [], // just 1 day so average doesn't make sense
+      GOMentionsCount: [], // covered in the total
       GOMentionsTotal: (mentionsToShow.length > 0) ? mentionsToShow : [],
-      GOMentionsAverage: [],
-      GOMentionsExclude: [],
+      GOMentionsAverage: [], // just 1 day so average doesn't make sense
       GOChecklistRefNote: "",
     }
 
@@ -194,8 +185,8 @@ export async function makeTodayProgress(itemsToShowArr: Array<string> = [], sour
       logError('makeTodayProgress', `Cannot find daily note to write to`)
     }
     return ''
-  } catch (error) {
-    logError('makeTodayProgress', error.message)
+  } catch (err) {
+    logError('makeTodayProgress', `Error: ${err.message}`)
     return '<error>' // for completeness
   }
 }